--- a/src/Hagar.ISerializable/ISerializableSerializer.cs
+++ b/src/Hagar.ISerializable/ISerializableSerializer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Buffers;
 using Hagar.Buffers;
 using Hagar.Session;
 
@@ -6,7 +7,7 @@ namespace Hagar.ISerializable
 {
     internal interface ISerializableSerializer
     {
-        void WriteValue(ref Writer writer, SerializerSession session, object value);
+        void WriteValue<TBufferWriter>(ref Writer<TBufferWriter> writer, SerializerSession session, object value) where TBufferWriter : IBufferWriter<byte>;
         object ReadValue(ref Reader reader, SerializerSession session, Type type, uint placeholderReferenceId);
     }
 }