… src/Hagar/Codecs/FieldHeaderCodec.cs
… src/Hagar/Codecs/IntegerCodec.cs
… src/Hagar/Codecs/KeyValuePairCodec.cs
… src/Hagar/Codecs/ListCodec.cs
… src/Hagar/Codecs/MultiDimensionalArrayCodec.cs
… src/Hagar/Codecs/ObjectCodec.cs
… src/Hagar/Codecs/ReferenceCodec.cs
… src/Hagar/Codecs/SkipFieldExtension.cs
… src/Hagar/Codecs/StringCodec.cs
… src/Hagar/Codecs/TagDelimitedFieldCodec.cs
… src/Hagar/Codecs/TupleCodec.cs
… src/Hagar/Codecs/TypeSerializerCodec.cs
… src/Hagar/Codecs/ValueTupleCodec.cs
… src/Hagar/Codecs/VoidCodec.cs
… src/Hagar/Configuration/SerializerConfiguration.cs
… src/Hagar/GeneratedCodeHelpers/HagarGeneratedCodeHelper.cs
… src/Hagar/Serializer.cs
… src/Hagar/Serializers/AbstractTypeSerializer.cs
… src/Hagar/Serializers/CodecProvider.cs
… src/Hagar/Serializers/ConcreteTypeSerializer.cs
… src/Hagar/Serializers/IPartialSerializer.cs
… src/Hagar/Serializers/IValueSerializer.cs
… src/Hagar/Serializers/ValueSerializer.cs
… src/Hagar/ServiceProviderExtensions.cs
… src/Hagar/Session/ReferencedObjectCollection.cs
… src/Hagar/Session/SessionPool.cs
… src/Hagar/TypeSystem/CachedTypeResolver.cs
… src/Hagar/TypeSystem/TypeCodec.cs
… src/Hagar/Utilities/CachedReadConcurrentDictionary.cs
… src/Hagar/Utilities/PrefixVarintCodec.cs