Optimized serialization path, mostly. I will look at deserialization in the future (maybe in a month or so).

SerializeBenchmark:

|            Method |        Mean |     Error |    StdDev | Scaled | ScaledSD |  Gen 0 | Allocated | Payload |
|------------------ |------------:|----------:|----------:|-------:|---------:|-------:|----------:|--------:|
|             Hagar |   136.08 ns |  2.364 ns |  2.211 ns |   1.00 |     0.00 |      - |       0 B |    20 B |
|           Orleans | 1,992.91 ns | 46.880 ns | 52.107 ns |  14.65 |     0.44 | 0.2403 |    1016 B |    77 B |
| MessagePackCSharp |    80.91 ns |  1.595 ns |  1.899 ns |   0.59 |     0.02 | 0.0094 |      40 B |    10 B |
|       ProtobufNet |   584.05 ns | 11.447 ns | 10.147 ns |   4.29 |     0.10 | 0.0496 |     208 B |    18 B |
|          Hyperion |   175.56 ns |  3.285 ns |  2.743 ns |   1.29 |     0.03 | 0.0226 |      96 B |    39 B |
|    NewtonsoftJson | 1,979.60 ns | 41.020 ns | 80.006 ns |  14.55 |     0.62 | 0.4807 |    2032 B |   154 B |


DeserializeBenchmark:

|            Method |       Mean |     Error |    StdDev |     Median | Scaled | ScaledSD |  Gen 0 | Allocated |
|------------------ |-----------:|----------:|----------:|-----------:|-------:|---------:|-------:|----------:|
|             Hagar |   590.5 ns | 11.710 ns | 17.526 ns |   585.7 ns |   1.00 |     0.00 | 0.0124 |      56 B |
|           Orleans | 1,250.9 ns | 24.601 ns | 26.322 ns | 1,252.7 ns |   2.12 |     0.08 | 0.1049 |     448 B |
| MessagePackCSharp |   103.5 ns |  2.085 ns |  5.744 ns |   102.7 ns |   0.18 |     0.01 | 0.0131 |      56 B |
|       ProtobufNet |   975.1 ns | 19.078 ns | 16.912 ns |   973.0 ns |   1.65 |     0.06 | 0.0267 |     112 B |
|          Hyperion |   219.9 ns |  4.306 ns |  6.954 ns |   218.3 ns |   0.37 |     0.02 | 0.0379 |     160 B |
|     ZeroFormatter |   354.0 ns |  7.080 ns | 17.631 ns |   347.9 ns |   0.60 |     0.03 | 0.0515 |     216 B |
|    NewtonsoftJson | 3,724.3 ns | 73.696 ns | 81.913 ns | 3,716.5 ns |   6.31 |     0.23 | 0.6790 |    2856 B |

StructSerializeBenchmark:

|            Method |        Mean |     Error |    StdDev | Scaled | ScaledSD |  Gen 0 | Allocated | Payload |
|------------------ |------------:|----------:|----------:|-------:|---------:|-------:|----------:|--------:|
|             Hagar |   129.09 ns |  2.577 ns |  3.259 ns |   1.00 |     0.00 |      - |       0 B |    20 B |
|           Orleans | 1,875.60 ns | 37.311 ns | 44.416 ns |  14.54 |     0.49 | 0.2518 |    1072 B |    78 B |
| MessagePackCSharp |    82.63 ns |  1.742 ns |  2.203 ns |   0.64 |     0.02 | 0.0094 |      40 B |    10 B |
|       ProtobufNet |   490.14 ns |  9.666 ns | 11.870 ns |   3.80 |     0.13 | 0.0625 |     264 B |     0 B |
|          Hyperion |   195.30 ns |  3.772 ns |  3.529 ns |   1.51 |     0.05 | 0.0360 |     152 B |    39 B |
|     ZeroFormatter |    99.47 ns |  1.843 ns |  1.724 ns |   0.77 |     0.02 | 0.0151 |      64 B |    36 B |
|    NewtonsoftJson | 1,828.97 ns | 32.833 ns | 30.712 ns |  14.18 |     0.42 | 0.4673 |    1968 B |   145 B |

StructDeserializeBenchmark:

|            Method |        Mean |     Error |    StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
|------------------ |------------:|----------:|----------:|-------:|---------:|-------:|----------:|
|             Hagar |   463.85 ns |  7.587 ns |  7.097 ns |   1.00 |     0.00 |      - |       0 B |
|           Orleans | 1,101.86 ns | 21.960 ns | 31.494 ns |   2.38 |     0.08 | 0.1049 |     448 B |
| MessagePackCSharp |    91.48 ns |  1.855 ns |  4.111 ns |   0.20 |     0.01 |      - |       0 B |
|       ProtobufNet | 1,031.81 ns | 20.441 ns | 27.289 ns |   2.22 |     0.07 | 0.0401 |     168 B |
|          Hyperion |   161.97 ns |  3.015 ns |  4.024 ns |   0.35 |     0.01 | 0.0379 |     160 B |
|     ZeroFormatter |    67.56 ns |  1.337 ns |  1.313 ns |   0.15 |     0.00 |      - |       0 B |
|    NewtonsoftJson | 5,405.07 ns | 95.542 ns | 89.370 ns |  11.66 |     0.25 | 0.8163 |    3432 B |